Change Id type in API to int64

+++ src/main/java/org/example/TTMS/responses/Trip.java
@@ -2,8 +2,8 @@ package org.example.TTMS.responses;
 
 import java.util.Map;
 
-public record Trip(int departureStationId, String departureStationName, String departureDateTime,
-                   int arrivalStationId, String arrivalStationName, String arrivalDateTime,
+public record Trip(long departureStationId, String departureStationName, String departureDateTime,
+                   long arrivalStationId, String arrivalStationName, String arrivalDateTime,
                    Map<String, String> price, String serviceName, String vehicleType,
                    boolean reservationAvailble, boolean reservationRequired) {
 }
